mux: Simplify handshakes
